+++ src/gfxlib2/win32/gfx_win32.c
@@ -757,9 +757,9 @@ void fb_hScreenInfo(int *width, int *height, int *depth, int *refresh)
 	ReleaseDC(NULL, hdc);
 }
 
-int fb_hGetWindowHandle(void)
+ssize_t fb_hGetWindowHandle(void)
 {
-	return (int)fb_win32.wnd;
+	return (ssize_t)fb_win32.wnd;
 }
 
 static void keyconv_clear( KEYCONVINFO *k )
