+++ src/rtlib/array_resetdesc.c
@@ -3,13 +3,9 @@
 
 #include "fb.h"
 
-/*:::::*/
-FBCALL void fb_ArrayResetDesc
-	( 
-		FBARRAY *array 
-	)
+FBCALL void fb_ArrayResetDesc( FBARRAY *array )
 {
-    int	i;
+	size_t i;
     FBARRAYDIM *dim;
 
     dim = &array->dimTB[0];
@@ -26,4 +22,3 @@ FBCALL void fb_ArrayResetDesc
 
     FB_ARRAY_SETDESC( array, array->element_len, 0, 0, 0 );
 }
-
