+++ src/rtlib/dev_file_size.c
@@ -5,7 +5,7 @@
 int fb_hDevFileSeekStart( FILE *fp, int mode, FB_FILE_ENCOD encod, int seek_zero )
 {
 	/* skip the BOM if in UTF-mode */
-	int ofs;
+	size_t ofs;
 
 	switch( encod )
 	{
