+++ src/rtlib/file_putarray.c
@@ -2,13 +2,11 @@
 
 #include "fb.h"
 
-/*:::::*/
-FBCALL int fb_FilePutArray( int fnum, long pos, FBARRAY *src )
+FBCALL int fb_FilePutArray( int fnum, int pos, FBARRAY *src )
 {
 	return fb_FilePutData( fnum, pos, src->ptr, src->size, TRUE, FALSE );
 }
 
-/*:::::*/
 FBCALL int fb_FilePutArrayLarge( int fnum, long long pos, FBARRAY *src )
 {
 	return fb_FilePutData( fnum, pos, src->ptr, src->size, TRUE, FALSE );
